factor out I/O routines into fastpath_io
